Add keystonev3 service to the catalog
This is one way of exposing the v3 API by way of the catalog. The
api-paste already exposes the v3 API, but it wasn't advertised anywhere.
Clients still won't automatically use the v3 API as nothing really looks
for this service name/type, however users who wish to manually make use
of the v3 API can discover it by way of the catalog.

This change also removes long form reference to the auth_uri and
instead makes use of the variable, which will help if we ever change
versions exposed.